Added extra API Functions; Minor design changes to database

+++ DAL/Entities/A_Permission.cs
@@ -8,9 +8,9 @@ namespace DAL.Entities
 {
     public class A_Permission
     {
-        public int A_Permission_ID { get; set; }
-        public int Album_ID { get; set; }
-        public int User_ID { get; set; }
+        public Int64 A_Permission_ID { get; set; }
+        public Int64 Album_ID { get; set; }
+        public Int64 User_ID { get; set; }
         public String? A_Permission_Type { get; set; }
 
         public Album? Album { get; set; } //BN comment - why?
