+++ DAL/Entities/Album.cs
@@ -8,7 +8,7 @@ namespace DAL.Entities
 {
     public class Album
     {
-        public int Album_ID { get; set; }
+        public Int64 Album_ID { get; set; }
         public String? Album_Name { get; set; }
         public String? Album_Description { get; set; }
