+++ Gallery2/Models/I_Permission/I_PermissionUpdate_Pass_Object.cs
@@ -2,6 +2,6 @@
 {
     public class I_PermissionUpdate_Pass_Object : I_Permission_Pass_Object
     {
-        public int i_permission_id { get; set; }
+        public Int64 i_permission_id { get; set; }
     }
 }
