+++ LOGIC/Services/Models/Album/Album_ResultSet.cs
@@ -8,7 +8,7 @@ namespace LOGIC.Services.Models.Album
 {
     public class Album_ResultSet
     {
-        public int album_id { get; set; }
+        public Int64 album_id { get; set; }
         public String album_name { get; set; }
         public String album_description { get; set; }
        // public ICollection<Image> images { get; set; }
