+++ LOGIC/Services/Models/I_Permission/I_Permission_ResultSet.cs
@@ -8,9 +8,9 @@ namespace LOGIC.Services.Models.I_Permission
 {
     public class I_Permission_ResultSet
     {
-        public int i_permission_id { get; set; }
-        public int image_id { get; set; }
-        public int user_id { get; set; }
+        public Int64 i_permission_id { get; set; }
+        public Int64 image_id { get; set; }
+        public Int64 user_id { get; set; }
         public String i_permission_type { get; set; }
 
        // public Image Image { get; set; } //BN Comment - why add this??
